+++ src/js/index.js
@@ -1,2 +1,3 @@
 import './utils.js';
+import './common.js';
 import '../css/style.css';
