+++ src/js/utils.js
@@ -10,45 +10,6 @@
   switchViewport();
 })();
 
-/* global Splide */
-const setSlider = () => {
-  const splideOptions = {
-    arrows: false,
-    pagination: false,
-    autoWidth: true,
-    type: 'loop',
-    drag: false,
-    pauseOnHover: false,
-    pauseOnFocus: false,
-    autoScroll: {
-      speed: 0.4,
-      pauseOnHover: false,
-    },
-  };
-  const sliderIds = ['#slider1', '#slider2'];
-  sliderIds.forEach((id) => {
-    new Splide(id, splideOptions).mount(window.splide.Extensions);
-  });
-};
-
-const setCTA = () => {
-  if (window.matchMedia('(max-width: 767px)').matches) {
-    const cta = document.querySelector('.l-sticky-cta');
-    const top = document.querySelector('.l-top');
-
-    if (!cta || !top) {
-      return;
-    }
-    const topBottom = top.getBoundingClientRect().bottom + window.scrollY;
-
-    if (window.scrollY >= topBottom) {
-      cta.classList.add('l-sticky-cta--show');
-    } else {
-      cta.classList.remove('l-sticky-cta--show');
-    }
-  }
-};
-
 document.addEventListener('DOMContentLoaded', () => {
   const observerCallback = (entries) => {
     entries.forEach((entry) => {
@@ -71,22 +32,14 @@ document.addEventListener('DOMContentLoaded', () => {
       e.preventDefault();
       const target = document.querySelector(link.getAttribute('href'));
       if (target) {
-        target.scrollIntoView({ behavior: 'smooth' });
+        const header = document.querySelector('.l-header');
+        const headerHeight = header && window.innerWidth >= 768 ? header.offsetHeight : 0;
+        const targetPosition = target.getBoundingClientRect().top + window.scrollY - headerHeight;
+        window.scrollTo({
+          top: targetPosition,
+          behavior: 'smooth',
+        });
       }
     });
   });
-
-  setSlider();
-  setCTA();
-});
-
-window.addEventListener('load', () => {
-  const loadElms = document.querySelectorAll('.js-load');
-  loadElms.forEach((loadElm) => {
-    loadElm.classList.add('js-load--loaded');
-  });
-});
-
-window.addEventListener('scroll', () => {
-  setCTA();
 });
